refactor: ScrollToTop 메소드 폴더 이동
utils폴더로 이동
